rewrite the pure ruby example
This solves a bunch of problems while bringing additional benefits like
a more standard API and modern coding practice.

Existing issues with the 12 year old version:

1. Unclear usage of Bitset.  Is it the bitset gem?  If so, the code doesn't
   call the Bitset API properly.
2. Unnecessary usage of bitmask, which doubles storage requirements
3. Improper use of CRC32's "seed" value: passing serial numbers (0,1,2,etc)
   instead of the prior hash value
4. Unnecessary use of main() function with somewhat confusing contents
5. Unnecessary use of seed value

I fixed all of the above, and added a bunch of features:

1. Sensible and overrideable defaults
2. Made clear that the only valid items are Strings
3. Added BloomFilter#bits(str) which returns an array(num_hashes) of "on bits"
4. Conveniently, this is exactly what the Bitset API wants
5. Hence, drastically simpler and cheaper
6. Use more standard API (for Ruby): BloomFilter#add and BloomFilter#include?
7. Added BloomFilter#percent_full to track status of "on bits"
8. Added BloomFilter#fpr to track status of false percentage rate
9. Added BloomFilter#likelihood to return a number between 0.0 and 1.0
   where 0.0 is very common and means "definitely not present"
10. BloomFilter#to_s to see the status of the filter
11. Added aliases to enable the following
    bf = BloomFilter.new
    bf << 'foo'
    bf['foo'] #=> 1.0 # current FPR is below epsilon
    bf['bar'] #=> 0.0

    bf << 'bar'
    bf['bar'] #=> 1.0 # current FPR still below epsilon
    bf['baz'] #=> 0.0
12. Aliased #inspect to #to_s
13. Rewrote main() to use if __FILE__ == $0
14. Rewrote main() to read from $stdin, such that:
      cat examples/pure-ruby-bf.rb | ruby examples/pure-ruby-bf.rb
    gives some useful output
15. Reading from STDIN is now supported properly

Note that this is based on some work I've done that will be incorporated into
my compsci gem.  You may want to just point to that at some point, as I have
tests and a bunch of utilities for optimal sizing, particularly for trading
less hashing for slightly more bits.

As I write this commit message, I am working on integrating my main effort
into the compsci repo and gem: https://github.com/rickhull/compsci

I used the 12 year old version here as a starting point, so this is my way
of giving back.